refactor audio i/o functions into separate madmom.io.audio module
