+++ madmom/io/__init__.py
@@ -1,0 +1,9 @@
+# encoding: utf-8
+"""
+Input/output package.
+
+"""
+
+from __future__ import absolute_import, division, print_function
+
+from .audio import load_audio_file
